Resolve implicit conversion bug

+++ test/format/ops.c
@@ -1131,7 +1131,7 @@ rollback_retry:
             WT_ORDERED_READ(max_rows, table->rows_current);
         tinfo->keyno = mmrand(&tinfo->data_rnd, 1, (u_int)max_rows);
         if (TV(OPS_PARETO))
-            tinfo->keyno = testutil_pareto(tinfo->keyno, 1, (u_int)max_rows, 20);
+            tinfo->keyno = testutil_pareto((uint32_t)tinfo->keyno, 1, (u_int)max_rows, 20);
         replay_adjust_key(tinfo, max_rows);
 
         /*
